add disable_p2p flag

+++ voltaire_bundler/utils/SignalHaltError.py
@@ -20,7 +20,8 @@ class SignalHaltError(SystemExit):
         return f"\nExitted due to {self.signal_enum.name}"
 
 
-def immediate_exit(signal_enum: Signals, loop: AbstractEventLoop, p2p:Popen) -> None:
-    p2p.terminate()
+def immediate_exit(signal_enum: Signals, loop: AbstractEventLoop, p2p:Popen|None) -> None:
+    if p2p is not None:
+        p2p.terminate()
     loop.stop()
     raise SignalHaltError(signal_enum=signal_enum)
